allow non-logged-in users to define tag filters
when not logged in, store the filters in a long-lasting cookie and
do not cache the home page

for that one guy on hacker news that complained about lobste.rs not
having this

+++ app/controllers/filters_controller.rb
@@ -1,11 +1,19 @@
 class FiltersController < ApplicationController
-  before_filter :require_logged_in_user
+  before_filter :authenticate_user
 
   def index
     @cur_url = "/filters"
     @title = "Filtered Tags"
 
-    @filtered_tags = @user.tag_filters.reload
+    if @user
+      @filtered_tags = @user.tag_filters.reload
+    else
+      @filtered_tags = tags_filtered_by_cookie.map{|t|
+        tf = TagFilter.new
+        tf.tag = t
+        tf
+      }
+    end
 
     render :action => "index"
   end
@@ -20,22 +28,26 @@ class FiltersController < ApplicationController
       end
     end
 
-    @user.tag_filters(:include => :tag).each do |tf|
-      if tf.tag && new_filters.include?(tf.tag.tag)
-        new_filters.reject!{|t| t == tf.tag.tag }
-      else
-        tf.destroy
+    if @user
+      @user.tag_filters(:include => :tag).each do |tf|
+        if tf.tag && new_filters.include?(tf.tag.tag)
+          new_filters.reject!{|t| t == tf.tag.tag }
+        else
+          tf.destroy
+        end
       end
+
+      new_filters.each do |t|
+        tf = TagFilter.new
+        tf.user_id = @user.id
+        tf.tag_id = Tag.find_by_tag(t).id
+        tf.save!
+      end
+    else
+      cookies.permanent[TAG_FILTER_COOKIE] = new_filters.join(",")
     end
 
-    new_filters.each do |t|
-      tf = TagFilter.new
-      tf.user_id = @user.id
-      tf.tag_id = Tag.find_by_tag(t).id
-      tf.save
-    end
-
-    flash.now[:success] = "Your filters have been updated."
-    index
+    flash[:success] = "Your filters have been updated."
+    return redirect_to "/filters"
   end
 end
